v2: Remove certificate support
Dotege is architected rather weirdly and is in need of a good
refactor. AFAIA the certificate functionality isn't used by
any current deployments of Dotege, and adds significant
complexity. (Plus, it's a bit weird: why do we deal with
templating AND certificates?)

As this is a very major change, this marks the start of Dotege
v2.