v2: Remove support for users/acls
This is extremely niche, and awkward to configure. In the future
we'll allow generic info to be passed on to the template, so it
can write ACLs if required, or activate middlewares, or whatever.

For now, just drop all the functionality so the project can be
cleaned up more easily.